+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Added time tagging to channels
+- Added `Pulse.play()` method
 
 ### Fixed
 - Change location of port feedforward and feedback filters in config